Move API files to a src folder
