+++ api/src/stampy_chat/get_blocks.py
