+++ web/src/pages/index.tsx
@@ -1,5 +1,3 @@
-const API_URL = process.env.NEXT_PUBLIC_API_URL || "http://127.0.0.1:3000";
-
 import Head from "next/head";
 import React from "react";
 import { type NextPage } from "next";
@@ -11,6 +9,7 @@ import Header from "../header";
 import { SearchBox, Followup } from "../searchbox";
 import logo from "../logo.svg"
 import { GlossarySpan } from "../glossary";
+import { API_URL } from "../settings"
 
 type Citation = {
   title: string;
