+++ web/src/settings.ts
@@ -1,0 +1,2 @@
+export const API_URL = process.env.NEXT_PUBLIC_API_URL || "http://127.0.0.1:3001";
+export const STAMPY_URL = process.env.STAMPY_URL || "https://aisafety.info"
